requirements.txt
# Python 3.12+ required
h5py~=3.12.0
numpy~=2.2.2
optuna~=4.1.0
pillow~=11.1.0
rich~=14.0.0
scikit-learn~=1.5.2
scipy~=1.15.1
timm~=1.0.15
torch~=2.6.0
torchvision~=0.21.0
tqdm~=4.67.1
wandb~=0.19.6
lightning~=2.5.0
pyyaml~=6.0.1  # For YAML config support
pytest~=7.4.0  # For running unit tests